create diff.go from testcase
